pyproject.toml
[project]
name = "visa-mcp-server"
version = "1.0.0"
description = "MCP Server for Visa APIs - Foreign Exchange, ATM Locator, Subscription Manager, Stop Payment Service"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dedalus-mcp>=0.1.0",
    "dedalus-labs>=0.1.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
